Create CRUD for teachers

+++ app/Http/Controllers/TeacherController.php
@@ -2,8 +2,14 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\User;
+use App\Models\Gender;
+use App\Models\Subject;
 use App\Models\Teacher;
+use App\Models\Classroom;
+use App\Models\Nationality;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\DB;
 use Illuminate\Validation\Rules\Password;
 
 class TeacherController extends Controller
@@ -13,10 +19,32 @@ class TeacherController extends Controller
      */
     public function index()
     {
-        return response()->json([
-            'success' => true,
-            'teachers' => Teacher::all(),
-        ]);
+        $teachers = User::role('teacher')->paginate(15);
+        return view('admin.teachers', ['teachers' => $teachers]);
+        // return response()->json([
+        //     'success' => true,
+        //     'teachers' => Teacher::all(),
+        // ]);
+    }
+
+    /**
+     * Show the form for creating a new resource.
+     */
+    public function create()
+    {
+        $genders = Gender::all();
+        $nationalities = Nationality::all();
+        $classrooms = Classroom::all();
+        $subjects = Subject::all();
+        return view(
+            'partials.teacher-form',
+            [
+                'genders' => $genders,
+                'nationalities' => $nationalities,
+                'classrooms' => $classrooms,
+                'subjects' => $subjects,
+            ]
+        );
     }
 
     /**
@@ -27,39 +55,81 @@ class TeacherController extends Controller
         $validated = $request->validate([
             'first_name' => 'required|string|max:20',
             'last_name' => 'required|string|max:20',
-            "father's_name" => 'required|string|max:30',
-            "mother's_name" => 'required|string|max:30',
-            'phone_number' => 'required|string|max:10',
+            "father_name" => 'required|string|max:30',
+            "mother_name" => 'required|string|max:30',
+            'phone_number' => 'required|string|max:15',
             'address' => 'required|string|max:20',
             'date_of_birth' => 'required|date',
             'email' => 'required|email',
             'password' => Password::min(8)
-                            ->letters()
-                            ->mixedCase()
-                            ->numbers()
-                            ->symbols()
-                            ->uncompromised(3),
+                ->letters()
+                ->mixedCase()
+                ->numbers()
+                ->symbols()
+                ->uncompromised(3),
             'gender_id' => 'required|integer|max_digits:1',
             'nationality_id' => 'required|integer|max_digits:2',
+            'classroom_ids' => 'required|array',
+            'classroom_ids.*' => 'integer|exists:classrooms,id',
+            'subject_ids' => 'required|array',
+            'subject_ids.*' => 'integer|exists:subjects,id',
         ]);
         $validated['password'] = bcrypt($validated['password']);
-        $teacher = Teacher::create($validated);
-        return response()->json([
-            'success' => true,
-            'message' => 'Teacher created successfully',
-            'teacher' => $teacher,
-        ]);
+        $classrooms_ids = $validated['classroom_ids'];
+        sort($classrooms_ids);
+        unset($validated['classroom_ids']);
+        $subjects_ids = $validated['subject_ids'];
+        sort($subjects_ids);
+        unset($validated['subject_ids']);
+        $teacher = User::create($validated);
+        $teacher->assignRole('teacher');
+        foreach ($classrooms_ids as $classroom_id) {
+            DB::table('classroom_teacher')->insert([
+                'classroom_id' => $classroom_id,
+                'teacher_id' => $teacher->id,
+            ]);
+        }
+        foreach ($subjects_ids as $subject_id) {
+            DB::table('subject_teacher')->insert([
+                'subject_id' => $subject_id,
+                'teacher_id' => $teacher->id,
+            ]);
+        }
+        return to_route('teachers.index');
+        // return response()->json([
+        //     'success' => true,
+        //     'message' => 'Teacher created successfully',
+        //     'teacher' => $teacher,
+        // ]);
     }
 
     /**
      * Display the specified resource.
      */
-    public function show(Teacher $teacher)
+    public function show(User $teacher)
     {
-        return response()->json([
-            'success' => true,
-            'student' => $teacher,
-        ]);
+        $class_ids = DB::table('classroom_teacher')
+            ->where('teacher_id', '=', $teacher->id)
+            ->pluck('classroom_id');
+        $classes = DB::table('classrooms')
+            ->whereIn('id', $class_ids)
+            ->pluck('room_number')
+            ->toArray();
+        $subject_ids = DB::table('subject_teacher')
+            ->where('teacher_id', '=', $teacher->id)
+            ->pluck('subject_id');
+        $subjects = DB::table('subjects')
+            ->whereIn('id', $subject_ids)
+            ->pluck('subject')
+            ->toArray();
+        return view(
+            'admin.teacher',
+            [
+                'teacher' => $teacher,
+                'classes' => $classes,
+                'subjects' => $subjects,
+            ]
+        );
     }
 
     /**
@@ -77,11 +147,11 @@ class TeacherController extends Controller
             'date_of_birth' => 'required|date',
             'email' => 'required|email',
             'password' => Password::min(8)
-                            ->letters()
-                            ->mixedCase()
-                            ->numbers()
-                            ->symbols()
-                            ->uncompromised(3),
+                ->letters()
+                ->mixedCase()
+                ->numbers()
+                ->symbols()
+                ->uncompromised(3),
             'gender_id' => 'required|integer|max_digits:1',
             'nationality_id' => 'required|integer|max_digits:2',
         ]);
@@ -97,12 +167,13 @@ class TeacherController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(Teacher $teacher)
+    public function destroy(User $teacher)
     {
-        $teacher->delete();
-        return response()->json([
-            'success' => true,
-            'message' => 'Teacher deleted successfully',
-        ]);
+        if(User::role('teacher')) {
+            $teacher->delete();
+            return to_route('teachers.index')->with(['delete-successfully' => 'Teacher deleted successfully']);
+        } else {
+            return redirect()->back();
+        }
     }
 }
